The dimension of a weyl object

The dimension of a weyl algebra is the number of variables needed; it is half the spray::arity(). The dimension of a Weyl algebra generated by \(\left\lbrace x_1,x_2,\ldots,x_n,\partial_{x_1},\partial_{x_2},\ldots,\partial_{x_n}\right\rbrace\) is \(n\) (not \(2n\)).

Usage

# S3 method for class 'weyl'
dim(x)

Arguments

x

Object of class weyl

Value

Integer

Note

Empty spray objects give zero-dimensional weyl objects.

Author

Robin K. S. Hankin

Examples

(a <- rweyl())
#> A member of the Weyl algebra:
#>   x  y  z dx dy dz     val
#>   1  2  1  0  0  0  =    3
#>   1  0  0  0  2  2  =    2
#>   1  0  1  2  2  1  =    1
dim(a)
#> [1] 3